Fixing up unsetSum counting

+++ src/solver.h
@@ -314,7 +314,7 @@ class Solver : public Searcher
             vector<uint32_t> satisfies;
             vector<unsigned char> can_be_unset;
             vector<uint32_t>* trail_lim_vars;
-            uint32_t can_be_unsetSum;
+            int64_t can_be_unsetSum;
             bool must_fix;
             uint32_t num_fixed;
             bool verbose = false;
